+++ NEWS.md
@@ -1,3 +1,7 @@
+## MINOR CHANGES
+
+- Disallowed saving of missing values with non-missing flags (Fixes SWS-1411)
+
 ## BUG FIXES
 
 - `GetMetadata` now returns characters instead of factors